Move setup.py next to PARA_ATM directory
Setup.py needs to be next to the directory containing the Python
pacakge.  Currently the layout is a little awkward because it is also
next to NATS, but the alternative was to move all of the code into a
second subdirectory also called PARA_ATM.  Longer term, if NATS is
removed from this repo, having setup.py and the PARA_ATM package
directory in the root may make sense, if the sole purpose of the repo
is to house the PARA_ATM python package.