Allow centaur to be optional
Much of the PARA_ATM functionality does not need CENTAUR, so it makes
sense to enable PARA_ATM to run even when CENTAUR is not installed.
Previously it would not load due to PARA_ATM importing DataStore,
which imported centaur.  The workaround is to ignore centaur import
errors in this module.