Add basic unittest framework
Adding a single unittest, which can easily be expanded in the future.
For now it just tests that the LaunchApp.Container can be constructed,
but this will identify any exceptions that occur in the setup code.

If PARA_ATM is installed, the tests can be run via:
python -m unittest
which can be issued in either the "src" or "PARA_ATM" directories.

If PARA_ATM has not been installed, believe that the tests can be run
explicitly by:
python -m PARA_ATM/tests/test_para_atm.py
from the "src" directory.