Rework docs directory
Creating a fresh layout with a minimal set of content to be updated
later.  Differences from before are that the build directory is not
added (it is a build artifact), the build_sphinx.sh script is not
needed, flask_doc.py has not been added, index.rst and conf.py are in
the base directory (which follows the usual layout), and a README file
has been added with simple instructions on how to build and view the
documentation.