Add basic database commands and tests
Add new database commands addTable, readTable, and dropTable.  The add
and read functions are generic and don't assume any particular
format.

Enable autocommit property of the psycopg connection.  The
documentation recommends this for "long lived scripts" so that the
session doesn't remain "idle in transaction".

Add basic test cases for the database to verify ability to check if
table exists, add a table, and read a table.