Adding new read function read_iff_file
This is to replace the current code under Commands/readIFF.py.  The
changes are intended to simplify the code, speed it up, make it more
robust, and improve it to read any and all record types.

The new approach works by first reading the entire file in as a list
of lines, identifying the record type associated with each line, and
then reading the collection of lines for each record type separately
into a DataFrame.  Even for large ~150MB files, this takes only a few
seconds, so the added complexity of parallel processing is no longer
necessary.

A mapping of record types -> column labels is defined for a base IFF
format 2.6.  In some cases, newer versions add additional columns.
Using the old column specification would still work (and the new
columns would not be read), but some logic has been added to try and
account for some of the changes in versions 2.13 and 2.15, so that the
new columns are processed as well.

The read_iff_file function allows the user to request a single record
type, a list of record types, or 'all'.  The result will either be a
single DataFrame or a dictionary mapping record types to DataFrames.

The interpolation option has not been included (and it does not appear
that the current interpolation in readIFF.py is operational anyway).
Thinking that it may be better to place interpolation into a utility
function that can be shared across the package and that operates
separately from the read functions.

This version does not add a "status" column like the previous
version.  Suggest that if needed, utility functions could be written
to compute a status and provide more control over the specifics.
Prefer to limit the reader code to just reading.