Switch para_atm entry point to general-purpose command-line utility
Previously, para_atm was hard-coded to call LaunchApp.main.  Now, it
calls a script that handles command line arguments.  The intent is to
support various commands that can provide different functions.  For
now, the available commands are "app" and "plot".  The "app" command
will launch the GUI application as before.  The plot command is used
with a specified file name to create a trajectory plot of the data in
the given file (NATS or IFF data).

For the updates to take effect, the package needs to be re-installed.