Further memory improvements to read_iff_file
Instead of grabbing a subset of lines, joining them, putting them into
a StringIO object, and reading from that, now we just figure out which
lines to skip and pass that to pd.read_csv via the skip_rows option.
This way the file data is never duplicated in memory.  Previously, the
file data may have been in memory 3 times (one for the list of lines,
a second for the joined version in the StringIO option, and a third
one for the DataFrame being created).

In informal testing with 150MB file, this did seem to reduce memory
usage during the read operation by about a factor of 2-3.