Move db access into Container class in LaunchApp
Otherwise, any system in which the database was not yet set up would
not be able to import the LaunchApp module, which meant that the test
file would fail to load and the para_atm command would not run (even
"para_atm -h" would fail).

With this update, the testing can be run and the the para_atm command
can be used, even if the database has not been setup (of course, the
database-related test cases will fail, and "para_atm app" will not
work).