Adding NATS simulation test case
Test case is based on NATS DEMO_Gate_To_Gate_Simulation_SFO_PHX_beta1,
but it is implemented using the new NatsSimulationWrapper.  Note that
this test will only run if NATS is installed and the NATS_HOME
environment variable is set appropriately (NATS may be anywhere on the
system, as path issues are handled behind the scenes).  The test case
simply runs the simulation and verifies that the resulting DataFrame
has the expected number of rows.