Rework of how NATS JVM wrapping is done to support multiple runs
In the previous approach, the JVM was shut down after a simulation
completed, but that made it impossible to run future simulations in
the same Python process, because the JVM cannot be restarted.

To address this, the functions for starting and stopping the JVM have
been moved outside of the simulation wrapper into a new utility class,
NatsEnvironment.  This class provides static methods to start and stop
the JVM, as well as the NATS standalone server.  It takes care to
ensure that the JVM is not started more than once and that it is shut
down at exit.